cleaning up for PR
