+++ .vscode/launch.json
@@ -14,8 +14,6 @@
         "--colors",
         "--timeout",
         "999999",
-        "-g",
-        "WIP",
         "${workspaceFolder}/lib/test/libraryTest/**/*.js"
       ],
       "preLaunchTask": "watch",
