+++ src/parser/nodeIdMap/nodeIdMapUtils.ts
@@ -528,34 +528,8 @@ export function hasParsedToken(nodeIdMapCollection: Collection, nodeId: number):
     return false;
 }
 
-// Recalculates the id numbers of the Ast, starting with the given TXorNode and continuing for all of its children.
-// Used to help reset a node's id for recursive node kinds, such as RecursivePrimaryExpression.
-//
-// Mutates the NodeIdMap.Collection and the TXorNodes it holds.
-// Assumes the given arguments are valid before as this function does no validation.
-export function recalculateId(nodeIdMapCollection: Collection, parserState: IParserState, nodeStart: TXorNode): void {
-    // A helper stack we use for recursively visiting children nodes.
-    const newNodeIdByOldNodeId: Map<number, number> = new Map();
-    let nodeStack: TXorNode[] = [nodeStart];
-    let currentNode: TXorNode | undefined = nodeStack.pop();
-    while (currentNode !== undefined) {
-        const newNodeId: number = ParseContextUtils.nextId(parserState.contextState);
-        newNodeIdByOldNodeId.set(currentNode.node.id, newNodeId);
-
-        const childrenOfCurrentNode: ReadonlyArray<TXorNode> = NodeIdMapIterator.expectXorChildren(
-            nodeIdMapCollection,
-            currentNode.node.id,
-        );
-        const reversedChildrenOfCurrentNode: ReadonlyArray<TXorNode> = [...childrenOfCurrentNode].reverse();
-        nodeStack = nodeStack.concat(reversedChildrenOfCurrentNode);
-
-        currentNode = nodeStack.pop();
-    }
-
-    updateNodeIds(nodeIdMapCollection, newNodeIdByOldNodeId);
-}
-
-// Given a mapping of (existingId) => (newId) thism utates the NodeIdMap.Collection and the TXorNodes it holds.
+// Given a mapping of (existingId) => (newId) this mutates the NodeIdMap.Collection and the TXorNodes it holds.
+// Assumes the given arguments are valid as this function does no validation.
 export function updateNodeIds(nodeIdMapCollection: Collection, newNodeIdByOldNodeId: Map<number, number>): void {
     // We'll be iterating over them twice, so grab them once.
     const xorNodes: ReadonlyArray<TXorNode> = NodeIdMapIterator.expectXorNodes(nodeIdMapCollection, [
